# shuf: use rand::distributions::Uniform instead of a handwritten distribution