shuf currently uses a handwritten, and needlessly inefficient, code to sample from a uniform distribution. This PR updates the code to use `rand::distributions::Uniform` instead.

There is probably no real world performance improvement, but this should make the code a bit clearer. The randomization usually isn't the bottleneck, but it's now about faster, e.g. `seq 10000 | time shuf -rn 200000000 > /dev/null` is ~2x faster.