@tertsdiepraam 2022-02-08
Hi! Sorry for not responding earlier. I liked this PR, but forgot what it did exactly when I implemented https://github.com/uutils/coreutils/pull/3099. I also ran some benchmarks on this PR and while it about 1.5-2 times faster when `-r` was passed, it was also about 2-2.5 times slower when `-r` wasn't passed. So, we'll probably go with #3099, but thank you for your contribution!

<details>
<summary>Click here for the detailed benchmarks </summary>

`shuf` is GNU, `./shuf` is this PR and `./shuf.old` is the `main` branch.

## Without `-r`: 2.27 times slower

```
Benchmark 1: ./shuf -i 0-10000000
  Time (mean ± σ):      2.289 s ±  0.035 s    [User: 2.115 s, System: 0.170 s]
  Range (min … max):    2.261 s …  2.381 s    10 runs
 
Benchmark 2: ./shuf.old -i 0-10000000
  Time (mean ± σ):      1.008 s ±  0.006 s    [User: 0.839 s, System: 0.168 s]
  Range (min … max):    1.000 s …  1.020 s    10 runs
 
Benchmark 3: shuf -i 0-10000000
  Time (mean ± σ):      1.268 s ±  0.018 s    [User: 1.237 s, System: 0.029 s]
  Range (min … max):    1.260 s …  1.319 s    10 runs
 
  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet PC without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.
 
Summary
  './shuf.old -i 0-10000000' ran
    1.26 ± 0.02 times faster than 'shuf -i 0-10000000'
    2.27 ± 0.04 times faster than './shuf -i 0-10000000'
```

## With `-r`: 1.66 times faster

```
Benchmark 1: ./shuf -r -n 10000000 -i 0-1000
  Time (mean ± σ):     101.9 ms ±   1.0 ms    [User: 98.2 ms, System: 3.7 ms]
  Range (min … max):   100.7 ms … 104.6 ms    28 runs
 
Benchmark 2: ./shuf.old -r -n 10000000 -i 0-1000
  Time (mean ± σ):     169.2 ms ±   0.9 ms    [User: 164.3 ms, System: 4.9 ms]
  Range (min … max):   167.9 ms … 171.6 ms    17 runs
 
Benchmark 3: shuf -r -n 10000000 -i 0-1000
  Time (mean ± σ):     777.1 ms ±   4.7 ms    [User: 775.1 ms, System: 1.8 ms]
  Range (min … max):   772.3 ms … 785.3 ms    10 runs
 
Summary
  './shuf -r -n 10000000 -i 0-1000' ran
    1.66 ± 0.02 times faster than './shuf.old -r -n 10000000 -i 0-1000'
    7.62 ± 0.08 times faster than 'shuf -r -n 10000000 -i 0-1000'
```

</details>