@sylvestre 2022-03-03
@tertsdiepraam should we close this PR then ?
@camel-cdr Cool stuff, sorry it hasn't been merged